refactor: 구매금액 입력뷰와 관련된 코드를 모아 정리
