+++ src/main/kotlin/view/buyinginput/BuyingInputParsedResult.kt
