+++ src/test/kotlin/view/buyinginput/BuyingInputParserTest.kt
@@ -1,9 +1,12 @@
-package view
+package view.buyinginput
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
+import view.BuyingInput
+import view.BuyingInputParsedResult
+import view.InvalidInput
 
 internal class BuyingInputParserTest {
     @ParameterizedTest
